refactor service furthur
