+++ service/cronjob/committer/internal/logic/constants.go
@@ -16,20 +16,18 @@ import (
 )
 
 type (
-	// tx
 	Tx       = tx.Tx
 	TxDetail = tx.TxDetail
-	// block
+
 	Block          = block.Block
 	BlockForCommit = blockForCommit.BlockForCommit
-	// mempool
+
 	MempoolTx       = mempool.MempoolTx
 	MempoolTxDetail = mempool.MempoolTxDetail
-	// assets
-	L2Nft = nft.L2Nft
-	// assets history
+
+	L2Nft        = nft.L2Nft
 	L2NftHistory = nft.L2NftHistory
-	// account history
+
 	Account        = account.Account
 	AccountHistory = account.AccountHistory
 
@@ -39,8 +37,8 @@ type (
 	Liquidity        = liquidity.Liquidity
 	LiquidityHistory = liquidity.LiquidityHistory
 
-	SysconfigModel = sysconfig.SysConfigModel
-	MempoolModel   = mempool.MempoolModel
+	SysConfigModel = sysconfig.SysConfigModel
+	MemPoolModel   = mempool.MemPoolModel
 	BlockModel     = block.BlockModel
 
 	L2NftModel        = nft.L2NftModel
@@ -50,9 +48,8 @@ type (
 )
 
 const (
-	// tx status
 	TxStatusPending = tx.StatusPending
-	// asset type
+
 	GeneralAssetType         = commonAsset.GeneralAssetType
 	LiquidityAssetType       = commonAsset.LiquidityAssetType
 	NftAssetType             = commonAsset.NftAssetType
@@ -79,7 +76,6 @@ const (
 )
 
 const (
-	// 15 minutes
 	MaxCommitterInterval = 60 * 1
 )
