+++ service/cronjob/witness/etc/config.yaml.example
@@ -1,4 +1,4 @@
-Name: witnessGenerator.cronjob
+Name: witness.cronjob
 
 Postgres:
   DataSource: host=127.0.0.1 user=postgres password=ZecreyProtocolDB@123 dbname=zkbas port=5432 sslmode=disable
